requirements.txt
# Core
PyMuPDF>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
beautifulsoup4>=4.12.0
pandas>=2.0.0
Pillow>=10.0.0
chardet>=5.0.0

# OCR – RapidOCR (ONNX-based, supports Chinese + English, lightweight)
rapidocr-onnxruntime>=1.4.0

# Resume parsing (LangChain + OpenAI)
langchain-core>=1.2.0
langchain-openai>=1.1.0
openai>=1.0.0
